backend/requirements.txt
fastapi
uvicorn
motor
pydantic
python-dotenv
python-multipart
passlib[bcrypt]
bcrypt==4.0.1
python-jose[cryptography]
email-validator